Improve git setup: forbid `main` commit and check for git hook first

+++ {{ cookiecutter.repo_name }}/bin/git_setup.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
 set -e
 
-
 # Configure git hook
 chmod +x .githooks/pre-commit
-ln -s $PWD/.githooks/pre-commit .git/hooks/pre-commit
+[ ! -f .git/hooks/pre-commit ] && ln -s $PWD/.githooks/pre-commit .git/hooks/pre-commit
